Extract some classes and modules
