+++ spec/weather_jp/day_weather_spec.rb
