+++ spec/weather_jp/weather_spec.rb
