+++ spec/weather_jp/wrapper_spec.rb
